Try to define the Naginata Kogaki (small kana) as a layer
If you always hold Qwerty Q first like a shift key, this should
feel the same for typing the kogaki. We might then use some of
the spare slots/combos for punctuation normally on the edit layers?

They may be a way to do this with the existing Naginata functions,
but thi just uses a ZMK macro for each kogaki.