Color Naginata layout based on kana usage
Based on Asahi newspaper corpus as reported in Tomaoka & Makioka (2004),
but doesn't seem to match my expectations. For example, why does あ (a)
get home-row right index when it is relatively low use (fourth vowel
after お (o) which is on the shifted layer)? While た (ta) was 7th and
is at Qwerty N (shared with お on shift)? I suspect the あ key is partly
done due to that key's heavy use for the ten-ten combo, so a lower use
letter on tap would reduce load on that key/finger.

And there could be errors in my calculations too. All in all, this just
raises more questions - and doesn't look great with the shifted kana in
green anyway.