Put tab on LH3 (most tucked left thumb key)
Doing this on all layers (including Naginata where tab
is useful for autocompletion within the IME, and the
Nav/Num layer where this is just &trans).

This key is much less comfy for my left hand than the
matching key on the right - to the point that I will
not miss it if switching to a 34-key 3x5_2 layout.

Tab here is below the 3-key bottom row combo for tab,
FDL in my layout (XCV in Qwerty).

I am thinking about putting Enter on the right below
the 3-key combo for Enter to match, perhaps as the tap
option with layer change on hold?