Build 34-key layout for Forager Acid too (drop most tucked thumbs)
Using the LAYER_FROM36 idea from Cem Aksoylar (creator of the Forager):
https://github.com/caksoylar/zmk-config

This assumes the middle thumb is the natural home key,
and the most comfortable (gets R and space), with this
and the most splayed key retained for the 34-key layout
dropping the most tucked thumb key (least comfortable).

Goes back to a single space thumb (happy with using either,
but combining shift and space with hold-tap was imperfect).
Will instead try holding space for navigation (which I
last tried with QMK/Vial defaults and didn't like).

This leaves the 3rd right thumb as a redundant plain &mo
for the num/nav layer. Might be nice for Enter, but will
take a little adjustment!

For Naginata, had to drop the ABC shift key which seemed
handy for katakana. The 3rd thumb key might be nice
here (shift, caps-lock, or caps word like behavior?).