Use balanced &lt with quick-tap-ms (off by default) for lots of spaces
My typing habits made hold-tap with shift and space a bad match (I did
not keep shift down until after releasing letters), but it seems to be
fine with layer change on hold and space on tap.

Had to set quick-tap-ms (off by default) to enable tap-and-hold for when
I want lots of spaces.